--- a/Caddyfile
+++ b/Caddyfile
@@ -1,13 +1,6 @@
 # Learn how to configure the Mercure.rocks Hub on https://mercure.rocks/docs/hub/config
 {
-    # Debug mode (disable it in production!)
-    {$DEBUG}
-    # HTTP/3 support
-    servers {
-        protocol {
-            experimental_http3
-        }
-    }
+	{$GLOBAL_OPTIONS}
 }
 
 {$SERVER_NAME:localhost}
@@ -15,20 +8,20 @@
 log
 
 route {
-    encode zstd gzip
+	encode zstd gzip
 
-    mercure {
-        # Transport to use (default to Bolt)
-        transport_url {$MERCURE_TRANSPORT_URL:bolt://mercure.db}
-        # Publisher JWT key
-        publisher_jwt {env.MERCURE_PUBLISHER_JWT_KEY} {env.MERCURE_PUBLISHER_JWT_ALG}
-        # Subscriber JWT key
-        subscriber_jwt {env.MERCURE_SUBSCRIBER_JWT_KEY} {env.MERCURE_SUBSCRIBER_JWT_ALG}
-        # Extra directives
-        {$MERCURE_EXTRA_DIRECTIVES}
-    }
+	mercure {
+		# Transport to use (default to Bolt)
+		transport_url {$MERCURE_TRANSPORT_URL:bolt://mercure.db}
+		# Publisher JWT key
+		publisher_jwt {env.MERCURE_PUBLISHER_JWT_KEY} {env.MERCURE_PUBLISHER_JWT_ALG}
+		# Subscriber JWT key
+		subscriber_jwt {env.MERCURE_SUBSCRIBER_JWT_KEY} {env.MERCURE_SUBSCRIBER_JWT_ALG}
+		# Extra directives
+		{$MERCURE_EXTRA_DIRECTIVES}
+	}
 
-    respond /healthz 200
+	respond /healthz 200
 
-    respond "Not Found" 404
+	respond "Not Found" 404
 }