--- a/caddy/go.mod
+++ b/caddy/go.mod
@@ -8,6 +8,7 @@ require (
 	github.com/caddyserver/caddy/v2 v2.4.3
 	github.com/caddyserver/certmagic v0.14.1 // indirect
 	github.com/dunglas/mercure v0.12.2
+	github.com/fsnotify/fsnotify v1.5.1 // indirect
 	github.com/google/uuid v1.3.0 // indirect
 	github.com/klauspost/cpuid/v2 v2.0.9 // indirect
 	github.com/miekg/dns v1.1.43 // indirect