--- a/config.go
+++ b/config.go
@@ -96,6 +96,8 @@ func SetFlags(fs *pflag.FlagSet, v *viper.Viper) {
 	fs.BoolP("use-forwarded-headers", "f", false, "enable headers forwarding")
 	fs.BoolP("demo", "D", false, "enable the demo mode")
 	fs.BoolP("subscriptions", "s", false, "dispatch updates when subscriptions are created or terminated")
+	fs.BoolP("perf-experiment", "yolo", false, "enable performance experiment for topic selector store cache")
+	fs.Int64P("topic-cache-shard-size", "tcsz", DefaultTopicSelectorStoreLruShardCount, "size of each shard in topic selector store cache")
 
 	fs.Bool("metrics-enabled", false, "enable metrics")
 	fs.String("metrics-addr", "127.0.0.1:9764", "metrics HTTP server address")
@@ -150,9 +152,17 @@ func NewHubFromViper(v *viper.Viper) (*Hub, error) { //nolint:funlen,gocognit
 		return nil, fmt.Errorf("unable to create logger: %w", err)
 	}
 
-	tss, err := NewTopicSelectorStore(TopicSelectorStoreDefaultCacheNumCounters, TopicSelectorStoreCacheMaxCost)
-	if err != nil {
-		return nil, err
+	var tss *TopicSelectorStore
+	if v.GetBool("perf-experiment") {
+		tss, err = NewTopicSelectorStoreLru(DefaultTopicSelectorStoreLruMaxEntriesPerShard, v.GetInt64("topic-cache-shard-size"))
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		tss, err = NewTopicSelectorStore(TopicSelectorStoreDefaultCacheNumCounters, TopicSelectorStoreCacheMaxCost)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	if t := v.GetString("transport_url"); t != "" {