--- a/config.go
+++ b/config.go
@@ -96,8 +96,8 @@ func SetFlags(fs *pflag.FlagSet, v *viper.Viper) {
 	fs.BoolP("use-forwarded-headers", "f", false, "enable headers forwarding")
 	fs.BoolP("demo", "D", false, "enable the demo mode")
 	fs.BoolP("subscriptions", "s", false, "dispatch updates when subscriptions are created or terminated")
-	fs.BoolP("perf-experiment", "yolo", false, "enable performance experiment for topic selector store cache")
-	fs.Int64P("topic-cache-shard-size", "tcsz", DefaultTopicSelectorStoreLruShardCount, "size of each shard in topic selector store cache")
+	fs.Bool("yolo", false, "enable performance experiment for topic selector store cache")
+	fs.Int64("tcsz", DefaultTopicSelectorStoreLruShardCount, "size of each shard in topic selector store cache")
 
 	fs.Bool("metrics-enabled", false, "enable metrics")
 	fs.String("metrics-addr", "127.0.0.1:9764", "metrics HTTP server address")