--- a/config.go
+++ b/config.go
@@ -96,7 +96,7 @@ func SetFlags(fs *pflag.FlagSet, v *viper.Viper) {
 	fs.BoolP("use-forwarded-headers", "f", false, "enable headers forwarding")
 	fs.BoolP("demo", "D", false, "enable the demo mode")
 	fs.BoolP("subscriptions", "s", false, "dispatch updates when subscriptions are created or terminated")
-	fs.Int64("tcsz", DefaultTopicSelectorStoreLruMaxEntriesPerShard, "size of each shard in topic selector store cache")
+	fs.Int64("tcsz", DefaultTopicSelectorStoreLRUMaxEntriesPerShard, "size of each shard in topic selector store cache")
 
 	fs.Bool("metrics-enabled", false, "enable metrics")
 	fs.String("metrics-addr", "127.0.0.1:9764", "metrics HTTP server address")
@@ -153,7 +153,7 @@ func NewHubFromViper(v *viper.Viper) (*Hub, error) { //nolint:funlen,gocognit
 
 	var tss *TopicSelectorStore
 	if v.GetInt64("tcsz") > 0 {
-		tss, err = NewTopicSelectorStoreLru(v.GetInt64("tcsz"), DefaultTopicSelectorStoreLruShardCount)
+		tss, err = NewTopicSelectorStoreLRU(v.GetInt64("tcsz"), DefaultTopicSelectorStoreLRUShardCount)
 		if err != nil {
 			return nil, err
 		}