--- a/subscribe.go
+++ b/subscribe.go
@@ -77,15 +77,16 @@ func (h *Hub) SubscribeHandler(w http.ResponseWriter, r *http.Request) {
 
 // registerSubscriber initializes the connection.
 func (h *Hub) registerSubscriber(w http.ResponseWriter, r *http.Request) *Subscriber {
-	s := NewSubscriber(retrieveLastEventID(r), h.logger, h.topicSelectorStore)
+	s := NewSubscriber(retrieveLastEventID(r), h.logger)
 	s.Debug = h.debug
 	s.RemoteAddr = r.RemoteAddr
+	var privateTopics []string
 
 	if h.subscriberJWT != nil {
 		claims, err := authorize(r, h.subscriberJWT, nil)
 		if claims != nil {
 			s.Claims = claims
-			s.TopicSelectors = claims.Mercure.Subscribe
+			privateTopics = claims.Mercure.Subscribe
 		}
 		if err != nil || (claims == nil && !h.anonymous) {
 			http.Error(w, http.StatusText(http.StatusUnauthorized), http.StatusUnauthorized)
@@ -97,14 +98,13 @@ func (h *Hub) registerSubscriber(w http.ResponseWriter, r *http.Request) *Subscr
 		}
 	}
 
-	s.Topics = r.URL.Query()["topic"]
-	if len(s.Topics) == 0 {
+	var topics = r.URL.Query()["topic"]
+	if len(topics) == 0 {
 		http.Error(w, "Missing \"topic\" parameter.", http.StatusBadRequest)
 
 		return nil
 	}
-	s.EscapedTopics = escapeTopics(s.Topics)
-	go s.start()
+	s.SetTopics(topics, privateTopics)
 
 	h.dispatchSubscriptionUpdate(s, true)
 	if err := h.transport.AddSubscriber(s); err != nil {