This PR is a draft. It's just here for discussion.   
This PR will be deleted and a new PR will be created against upstream once it is proven to fix the problem.