@divine 2021-09-12
Hello,

Well, first of all, thank you so much for taking a look and having this fixed.

This indeed fixes the problem I was having, it's now performing **amazingly fast** and without using much CPU, it's well under %8-%10.

However, there are some issues that I've observed.

If you don't have time, I'm ready to fix these issues and push PR towards this repository/branch - just let me know.

- Local transport doesn't deliver messages due to missing (`s.Ready()`):
https://github.com/kevburnsjr/mercure/blob/b302b59932abc33a27abfe8b8f12c2e5d6b1ff74/bolt_transport.go#L177-L196
in function https://github.com/kevburnsjr/mercure/blob/b302b59932abc33a27abfe8b8f12c2e5d6b1ff74/local_transport.go#L52-L68

- Caddy configuration still uses the old topic selector and it's not configurable:
https://github.com/kevburnsjr/mercure/blob/b302b59932abc33a27abfe8b8f12c2e5d6b1ff74/caddy/caddy.go#L139-L142

No other issues that I have found and again thank you so much!

Thanks!