@kevburnsjr 2021-09-13
Done and done. Tests are fixed to.

> amazingly fast and without using much CPU, it's well under %8-%10.

Got any graphs? 🚀