@divine 2021-09-13
> Done and done. Tests are fixed to.
> 
> > amazingly fast and without using much CPU, it's well under %8-%10.
> 
> Got any graphs? 🚀

Caddy file doesn't parse it, looks like it needs to parse algorithm after this case https://github.com/kevburnsjr/mercure/blob/subscriber-performance/caddy/caddy.go#L352:

```
case "lru_shard_size":
  if !d.NextArg() {
    return d.ArgErr()
  }
  
  v, err := strconv.ParseInt(d.Val(), 10, 64)
  if err != nil {
    return err //nolint:wrapcheck
  }
  
  m.LruShardSize = &v
```


Regarding graphs, do you mean `pprof` graphs, right? If yes, here we go:

![kevfix](https://user-images.githubusercontent.com/48183131/133008759-25a1e889-2257-4bfb-b9f9-730d3ff87dc6.png)


Again, thank you so much!