@divine 2021-09-13
Another graph after running for about 15 minutes:

```
# HELP mercure_subscribers The current number of running subscribers
# TYPE mercure_subscribers gauge
mercure_subscribers 2773
# HELP mercure_subscribers_total Total number of handled subscribers
# TYPE mercure_subscribers_total counter
mercure_subscribers_total 9155
# HELP mercure_updates_total Total number of handled updates
# TYPE mercure_updates_total counter
mercure_updates_total 64149

```

![kevfix2](https://user-images.githubusercontent.com/48183131/133009098-aa575523-adda-4dc3-b33e-577a1ca10dae.png)