@kevburnsjr 2021-09-13
@divine I actually meant newrelic graphs or grafana or whatever shows the CPU usage dropping off a cliff like an action film.

I see this in the tests. 

```
	{
		http_port     9080
		https_port    9443
	}
	localhost:9080 {
		route {
			mercure {
				anonymous
				subscriptions
				publisher_jwt !ChangeMe!
			}
	
			respond 404
		}
	}
```

I've never used or configured Caddy. I'm thinking a new directive:

```
			mercure {
				anonymous
				subscriptions
				publisher_jwt !ChangeMe!
				lru_cache 10000
			}
```

Makes sense to me. We can always change it when we switch to make lru the default and probably best to make it separate.