@divine 2021-09-13
@kevburnsjr sure, here we go:
Daily:
![image](https://user-images.githubusercontent.com/48183131/133009970-72a4ed94-50e1-41dd-a503-0e5529f3411e.png)
Weekly:
![image](https://user-images.githubusercontent.com/48183131/133009696-a2dd58f8-1d5d-4962-bf42-5fb514bbd13d.png)

![image](https://user-images.githubusercontent.com/48183131/133010131-d7227cc8-1c08-4fe1-b7f2-1e71d98cfeaf.png)

Yay, it's a really small virtual server that handles a few millions of notifications per day 😄.

> I'm thinking a new directive:

`lru_cache` makes sense! Maybe even shorter just `lru` but then it won't make sense probably.

---
JFYI, it's being used on https://mail.tm/en/ and currently deployed you can check how fast it is by sending email :)

Thanks!