@kevburnsjr 2021-09-14
@divine How do those numbers look? I was concerned they may have been creeping back up. Still looking good?