@divine 2021-09-14
Hello @kevburnsjr,

Well, after running for a day it was just stuck for no reason. I have double-checked the memory and CPU.

It was used both fully and notifications weren't dispatched to subscribed clients.

Probably the reason was a small server (EPYC 2 vCPU, 2GB RAM), I have spun up a new one (EPYC 3 vCPU, 4GB RAM) which looks a little better but still, CPU usage is a little bit high after running for a few hours.

I will report to you with my findings:

```
# HELP mercure_subscribers The current number of running subscribers
# TYPE mercure_subscribers gauge
mercure_subscribers 5154
# HELP mercure_subscribers_total Total number of handled subscribers
# TYPE mercure_subscribers_total counter
mercure_subscribers_total 726251
# HELP mercure_updates_total Total number of handled updates
# TYPE mercure_updates_total counter
mercure_updates_total 5.94848e+06
```

![image](https://user-images.githubusercontent.com/48183131/133278268-b3ceb334-a4e8-46aa-a9b7-feb2d05d76d8.png)

HTOP:
![image](https://user-images.githubusercontent.com/48183131/133278432-b509a35b-18dd-4c25-b51d-9da6f0dd290d.png)

Caddy configuration:
`lru_shard_size 500000`

pprof:

![kevfix3](https://user-images.githubusercontent.com/48183131/133279479-f2b5a55e-a00b-43dc-8498-3f4cf8c75b73.png)


[pprof.caddy.samples.cpu.030.pb.gz](https://github.com/kevburnsjr/mercure/files/7162908/pprof.caddy.samples.cpu.030.pb.gz)


Thanks!