@divine 2021-09-14
Increased the value of `lru_shard_size` to `1000000`

The result after the restart was fine then went up again due to high number of updates pushed to it:
![image](https://user-images.githubusercontent.com/48183131/133282262-19e01d32-0134-42b4-869d-2c5890913536.png)

Then went to normal back:


Almost idle:
![image](https://user-images.githubusercontent.com/48183131/133283938-aa644189-64a6-49c9-ae24-ba281a9b5abe.png)


![image](https://user-images.githubusercontent.com/48183131/133283999-de965d02-dc04-4132-a0d4-af56a7351f6a.png)

pprof:
![kevfix4](https://user-images.githubusercontent.com/48183131/133283310-d7580bda-5615-4876-b816-56ea31f3dab0.png)
[pprof.caddy.samples.cpu.031.pb.gz](https://github.com/kevburnsjr/mercure/files/7163043/pprof.caddy.samples.cpu.031.pb.gz)


```bash
# HELP mercure_subscribers The current number of running subscribers
# TYPE mercure_subscribers gauge
mercure_subscribers 5282
# HELP mercure_subscribers_total Total number of handled subscribers
# TYPE mercure_subscribers_total counter
mercure_subscribers_total 17811
# HELP mercure_updates_total Total number of handled updates
# TYPE mercure_updates_total counter
mercure_updates_total 33347
```

Thanks!