@kevburnsjr 2021-09-16
It's pretty cool that you can serve 5k subscribers on a 3 core VM. I though maybe to start looking for a memory leak but the fact that you've got CPU going up and down suggests it's probably not a leak.

The only leak I could find just by reading the code is here
https://github.com/dunglas/mercure/blob/main/subscribe.go#L107-L117

Where a subscriber is started, but the function returns nil so `Disconnect` is never called leaving the zombie subscription to sit and spin.
This error will always be seen in the logs as "Unable to add subscriber"
Do you see a lot of those?