@divine 2021-09-17
Hello!

> CPU going up and down suggests it's probably not a leak.

Exactly, it goes up and down when the high number of updates is pushed to it.

> Do you see a lot of those?

Actually, no such errors in the logs...

Thanks!