@kevburnsjr 2021-09-26
@divine I like mail.tm and I'd like to see it remain free. 
For that reason, I took another stab at improving efficiency. 

While the last set of enhancements was limited to improving the efficiency of the subscriber, this bit aims to improve the efficiency of message fan-out in the local transport. By caching a subscriber bitmap for each topic, we can reduce the number of cache lookups per update from `n` to `1` where `n` represents the number of subscribers. This new approach appears to be significantly more efficient for all workloads and can be over 100x more efficient for certain workloads. Care to give it a shot?

```
Before

cpu: Intel(R) Core(TM) i7-5820K CPU @ 3.30GHz
BenchmarkLocalTransport/topics_1_concurrency_100_matchpct_1-2               6020            197048 ns/op
BenchmarkLocalTransport/topics_1_concurrency_100_matchpct_10-2              6756            178345 ns/op
BenchmarkLocalTransport/topics_1_concurrency_100_matchpct_100-2             6175            185047 ns/op
BenchmarkLocalTransport/topics_1_concurrency_1000_matchpct_1-2               679           1774345 ns/op
BenchmarkLocalTransport/topics_1_concurrency_1000_matchpct_10-2              680           1753074 ns/op
BenchmarkLocalTransport/topics_1_concurrency_1000_matchpct_100-2             664           1801593 ns/op
BenchmarkLocalTransport/topics_1_concurrency_5000_matchpct_1-2               134           8964758 ns/op
BenchmarkLocalTransport/topics_1_concurrency_5000_matchpct_10-2              132           9114793 ns/op
BenchmarkLocalTransport/topics_1_concurrency_5000_matchpct_100-2             129           9199351 ns/op
BenchmarkLocalTransport/topics_1_concurrency_10000_matchpct_1-2               58          19172861 ns/op
BenchmarkLocalTransport/topics_1_concurrency_10000_matchpct_10-2              56          19190180 ns/op
BenchmarkLocalTransport/topics_1_concurrency_10000_matchpct_100-2             61          18511490 ns/op
BenchmarkLocalTransport/topics_5_concurrency_100_matchpct_1-2                331           3604988 ns/op
BenchmarkLocalTransport/topics_5_concurrency_100_matchpct_10-2               337           3533874 ns/op
BenchmarkLocalTransport/topics_5_concurrency_100_matchpct_100-2              606           1966361 ns/op
BenchmarkLocalTransport/topics_5_concurrency_1000_matchpct_1-2                32          35837120 ns/op
BenchmarkLocalTransport/topics_5_concurrency_1000_matchpct_10-2               33          35347438 ns/op
BenchmarkLocalTransport/topics_5_concurrency_1000_matchpct_100-2              57          19961933 ns/op
BenchmarkLocalTransport/topics_5_concurrency_5000_matchpct_1-2                 6         181751742 ns/op
BenchmarkLocalTransport/topics_5_concurrency_5000_matchpct_10-2                6         178616568 ns/op
BenchmarkLocalTransport/topics_5_concurrency_5000_matchpct_100-2              12          98426881 ns/op
BenchmarkLocalTransport/topics_5_concurrency_10000_matchpct_1-2                3         362162971 ns/op
BenchmarkLocalTransport/topics_5_concurrency_10000_matchpct_10-2               3         362948712 ns/op
BenchmarkLocalTransport/topics_5_concurrency_10000_matchpct_100-2              6         198618388 ns/op
BenchmarkLocalTransport/topics_10_concurrency_100_matchpct_1-2                82          14360454 ns/op
BenchmarkLocalTransport/topics_10_concurrency_100_matchpct_10-2               85          13955853 ns/op
BenchmarkLocalTransport/topics_10_concurrency_100_matchpct_100-2             141           8413187 ns/op
BenchmarkLocalTransport/topics_10_concurrency_1000_matchpct_1-2                8         143085475 ns/op
BenchmarkLocalTransport/topics_10_concurrency_1000_matchpct_10-2               8         140071410 ns/op
BenchmarkLocalTransport/topics_10_concurrency_1000_matchpct_100-2             13          84186038 ns/op
BenchmarkLocalTransport/topics_10_concurrency_5000_matchpct_1-2                2         723826206 ns/op
BenchmarkLocalTransport/topics_10_concurrency_5000_matchpct_10-2               2         693173880 ns/op
BenchmarkLocalTransport/topics_10_concurrency_5000_matchpct_100-2              3         418398172 ns/op
BenchmarkLocalTransport/topics_10_concurrency_10000_matchpct_1-2               1        1405078617 ns/op
BenchmarkLocalTransport/topics_10_concurrency_10000_matchpct_10-2              1        1375515850 ns/op
BenchmarkLocalTransport/topics_10_concurrency_10000_matchpct_100-2             2         846991015 ns/op
PASS
```
```
After

cpu: Intel(R) Core(TM) i7-5820K CPU @ 3.30GHz
BenchmarkLocalTransport/topics_1_concurrency_100_matchpct_1-2             264186              3947 ns/op
BenchmarkLocalTransport/topics_1_concurrency_100_matchpct_10-2            155938              7828 ns/op
BenchmarkLocalTransport/topics_1_concurrency_100_matchpct_100-2            25880             46677 ns/op
BenchmarkLocalTransport/topics_1_concurrency_1000_matchpct_1-2            148768              8461 ns/op
BenchmarkLocalTransport/topics_1_concurrency_1000_matchpct_10-2            25450             47901 ns/op
BenchmarkLocalTransport/topics_1_concurrency_1000_matchpct_100-2            3087            469505 ns/op
BenchmarkLocalTransport/topics_1_concurrency_5000_matchpct_1-2             40352             29910 ns/op
BenchmarkLocalTransport/topics_1_concurrency_5000_matchpct_10-2             8838            238562 ns/op
BenchmarkLocalTransport/topics_1_concurrency_5000_matchpct_100-2             546           2338087 ns/op
BenchmarkLocalTransport/topics_1_concurrency_10000_matchpct_1-2            21139             57396 ns/op
BenchmarkLocalTransport/topics_1_concurrency_10000_matchpct_10-2            3264            488230 ns/op
BenchmarkLocalTransport/topics_1_concurrency_10000_matchpct_100-2            262           5200914 ns/op
BenchmarkLocalTransport/topics_5_concurrency_100_matchpct_1-2             195138              6664 ns/op
BenchmarkLocalTransport/topics_5_concurrency_100_matchpct_10-2            118936             10112 ns/op
BenchmarkLocalTransport/topics_5_concurrency_100_matchpct_100-2            25681             47226 ns/op
BenchmarkLocalTransport/topics_5_concurrency_1000_matchpct_1-2            117304             10803 ns/op
BenchmarkLocalTransport/topics_5_concurrency_1000_matchpct_10-2            24054             48372 ns/op
BenchmarkLocalTransport/topics_5_concurrency_1000_matchpct_100-2            3770            469121 ns/op
BenchmarkLocalTransport/topics_5_concurrency_5000_matchpct_1-2             50228             30129 ns/op
BenchmarkLocalTransport/topics_5_concurrency_5000_matchpct_10-2             5155            235773 ns/op
BenchmarkLocalTransport/topics_5_concurrency_5000_matchpct_100-2             721           2422102 ns/op
BenchmarkLocalTransport/topics_5_concurrency_10000_matchpct_1-2            19496             61929 ns/op
BenchmarkLocalTransport/topics_5_concurrency_10000_matchpct_10-2            2468            481776 ns/op
BenchmarkLocalTransport/topics_5_concurrency_10000_matchpct_100-2            274           5617101 ns/op
BenchmarkLocalTransport/topics_10_concurrency_100_matchpct_1-2            118710             10484 ns/op
BenchmarkLocalTransport/topics_10_concurrency_100_matchpct_10-2            84920             13410 ns/op
BenchmarkLocalTransport/topics_10_concurrency_100_matchpct_100-2           24715             49139 ns/op
BenchmarkLocalTransport/topics_10_concurrency_1000_matchpct_1-2            87052             14284 ns/op
BenchmarkLocalTransport/topics_10_concurrency_1000_matchpct_10-2           23744             51394 ns/op
BenchmarkLocalTransport/topics_10_concurrency_1000_matchpct_100-2           2560            470790 ns/op
BenchmarkLocalTransport/topics_10_concurrency_5000_matchpct_1-2            36888             32576 ns/op
BenchmarkLocalTransport/topics_10_concurrency_5000_matchpct_10-2            5348            243063 ns/op
BenchmarkLocalTransport/topics_10_concurrency_5000_matchpct_100-2            505           2365070 ns/op
BenchmarkLocalTransport/topics_10_concurrency_10000_matchpct_1-2           19935             60491 ns/op
BenchmarkLocalTransport/topics_10_concurrency_10000_matchpct_10-2           2673            487560 ns/op
BenchmarkLocalTransport/topics_10_concurrency_10000_matchpct_100-2           200           5642333 ns/op
PASS
ok      github.com/dunglas/mercure      125.328s
```